+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 #### ⚡ Optimizations
 - Snapshot data can now be saved and loaded partially (e.g. only particle
   positions).
+- Multi-file GADGET snapshots can now be written in parallel.
 - Faster detrending of perturbations.
 
 #### 👌 Other changes